requirements.txt
ollama
mem0ai>=0.0.1
chromadb
python-dotenv
pydantic
